requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==4.2.14
django-cors-headers==4.4.0
django-ninja==1.2.1
numpy==2.0.0
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
